enable redis as store
